pyproject.toml
[project]
name = "multi-agent-analyst"
version = "0.1.0"
description = "A multi-agent CLI tool for automated data analysis"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Mudabbir" }]
dependencies = [
    "polars>=1.0.0",
    "typer[all]>=0.12.0",
    "rich>=13.0.0",
    "plotly>=5.20.0",
    "matplotlib>=3.8.0",
    "openpyxl>=3.1.0",
    "jinja2>=3.1.0",
    "kaleido>=0.2.1",
    "pantab>=3.0.0",
    "google-genai>=0.3.0",
    "xlsxwriter>=3.2.0",
    "questionary>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "python-multipart>=0.0.6",
    "pyjwt>=2.8.0",
    "cryptography>=42.0.0",
    "passlib[bcrypt]>=1.7.4"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
analyst = "analyst.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/analyst"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
